refactor: Kanban 관련 components 파일 계층구조화
- 칸반 컬럼과 칸반 모달을 제외한 것들은 다른 파일과 의존성이 없기 때문에 포함시킴
