+++ client/src/components/index.ts
@@ -15,10 +15,10 @@ export { default as EpicPlaceholer } from './EpicPlaceholder/index';
 
 // KanBan
 export { default as KanbanColumn } from './KanbanColumn/index';
-export { default as KanbanItem } from './KanbanItem/index';
-export { default as KanbanItemInput } from './KanbanItemInput/index';
-export { default as KanbanAddBtn } from './KanbanAddBtn/index';
-export { default as KanbanDeleteModal } from './KanbanDeleteModal/index';
+export { default as KanbanItem } from './KanbanColumn/KanbanItem/index';
+export { default as KanbanItemInput } from './KanbanColumn/KanbanItemInput/index';
+export { default as KanbanAddBtn } from './KanbanColumn/KanbanAddBtn/index';
+export { default as KanbanDeleteModal } from './KanbanColumn/KanbanDeleteModal/index';
 export { default as KanbanModal } from './KanbanModal/index';
 
 // Project
